+++ library/oci_database_facts.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright (c) 2018, Oracle and/or its affiliates.
+# Copyright (c) 2018, 2020 Oracle and/or its affiliates.
 # This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
@@ -211,8 +211,9 @@ def list_databases(db_client, module):
                 db_home_id=db_home_id,
             )
         else:
-            get_logger().error("Invalid parameter combination; need either database_id, or compartment_id and db_home_id")
-
+            get_logger().error(
+                "Invalid parameter combination; need either database_id, or compartment_id and db_home_id"
+            )
     except ServiceError as ex:
         get_logger().error("Unable to list Databases due to %s", ex.message)
         module.fail_json(msg=ex.message)